% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_envstat.R
\name{edit_envstat}
\alias{edit_envstat}
\title{Edit an envstat configuration file in RStudio}
\usage{
edit_envstat(filepath = "~/.envstat")
}
\arguments{
\item{filepath}{path to the config file}
}
\value{
the path to the config file invisibly
}
\description{
Edit an envstat configuration file in RStudio
}
