% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{sitrep}
\alias{sitrep}
\title{Output an environment status situsation report}
\usage{
sitrep(silent = FALSE, path = "~/.envstat")
}
\arguments{
\item{silent}{boolean disables printed output}

\item{path}{path to the envstat config file}
}
\value{
TRUE/FALSE invisibly, to indicate all checks pass/fail
}
\description{
Output an environment status situsation report
}
