% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_envstat.R
\name{use_envstat}
\alias{use_envstat}
\title{Create a new skeleton configuration file}
\usage{
use_envstat(path = "~/.envstat", source = NULL)
}
\arguments{
\item{path}{path to the config file}

\item{source}{customise the source of the example config file}
}
\value{
TRUE/FALSE invisibly to indicate success/failure
}
\description{
Create a new skeleton configuration file
}
