% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_envstat.R
\name{edit_envstat}
\alias{edit_envstat}
\title{Edit an envstat configuration file}
\usage{
edit_envstat(filepath = "~/.envstat")
}
\arguments{
\item{filepath}{path to the config file}
}
\value{
the path to the config file invisibly
}
\description{
Opens an \code{envstat} config file in your in your editor.
Supports either the RStudio IDE, or the R default editor from
\code{getOption("editor")}
}
\examples{
\dontrun{
# By default envstat uses a config file in your home directory
envstat::edit_envstat()

# But you can tell it to use a different config file if you prefer
envstat::edit_envstat(filepath = "/tmp/config.yml")
}
}
