% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{sitrep}
\alias{sitrep}
\title{Output an environment status situsation report}
\usage{
sitrep(path = "~/.envstat")
}
\arguments{
\item{path}{path to the envstat config file}
}
\value{
returns TRUE or FALSE invisibly, depending on whether all checks pass or fail
}
\description{
Output an environment status situsation report
}
